Add files via upload
This is a work in progress. The BlueBox is a 3D large volume printer based in the great Ultimaker printer, but more scalable (its also designed a Big BlueBox model) and easy to make by using aluminium profiles and easy to mount designed printed parts. Me and  Bernardo, good friend and collaborator in the BlueBox project, have a great lack of time, and the prototypes are not completely mounted. If you want to help in the development, or use the files to make your own fork, feel free. If you contact me, I`ll try to help you in any question related to this project.